#117 Update SimaSimulation
- clear code of SimaSimulation
- update test of SimaSimulation
